% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schedule}
\alias{schedule}
\title{schedule}
\format{An object of class \code{data.frame} with 7 rows and 5 columns.}
\usage{
schedule
}
\description{
Daniel Egan's twitter posts
}
\keyword{datasets}
