% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tweet_db}
\alias{tweet_db}
\title{tweet_db}
\format{An object of class \code{data.frame} with 28 rows and 4 columns.}
\usage{
tweet_db
}
\description{
Daniel Egan's twitter posts
}
\keyword{datasets}
